Agregue insert carga en el de documento

+++ Biblioteca/eliminar.php
@@ -1,30 +1,35 @@
 <?php
 include_once "../lib/PDOConfig.php";
 
-    if($_POST){
-    $id=$_POST['id'];
-    $ruta=$_POST['ruta'];
-    echo $ruta;
-    $base=new PDOConfig();
+if($_POST){
+  $id=$_POST['id'];
+  $ruta=$_POST['ruta'];
+  //  echo $ruta;
+  $base=new PDOConfig();
+  $sqlcarga="delete from carga where id_documento=".$id;
+  $resultadocarga=$base->query($sqlcarga);
+  if($resultadocarga){
     $sqlcontiene="delete from contiene where id_documento=".$id;
     $resultadocontiene=$base->query($sqlcontiene);
     if($resultadocontiene){
+      $sqldocumento="delete from documento where id_documento=".$id;
+      echo $sqldocumento;
+      $resultado=$base->query($sqldocumento);
+      if($resultado){
+        if(!unlink($ruta)){
+          echo "No se elimino el archivo";
+        }
+        echo "Exito al eliminar archivo";
 
-            $sqldocumento="delete from documento where id_documento=".$id;
-            $resultado=$base->query($sqldocumento);
-            if($resultado){
-                if(!unlink($ruta)){
-                  echo "No se elimino el archivo";
-                }
-                echo "Exito al eliminar archivo";
-
-            }else{
-                echo "Error al eliminar archivo";
-            }
+      }else{
+        echo "Error al eliminar archivo";
+      }
     }else{
-        echo "Error nose pudo eliminar archivo";
+      echo "Error no se pudo eliminar archivo";
     }
-
-    }
+  }else{
+    echo "Error no elimino carga";
+  }
+}
 
 ?>
